requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.1
cryptography==3.3.1
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.44.0
django-phonenumber-field==4.0.0
django-resized==0.3.9
django-webmaster-verification==0.4.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
newsapi-python==0.2.6
oauthlib==3.1.0
phonenumberslite==8.12.2
Pillow==8.0.1
pycparser==2.20
PyJWT==2.0.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2019.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.26.2
